+++ yash-cli/CHANGELOG.md
@@ -29,6 +29,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   the next command line. This behavior basically conforms to POSIX.1-2024, but
   differs in that the shell does not resume with the remaining commands
   following the next asynchronous and-or list.
+- When the shell starts job control, if it is in the background, the shell now
+  suspends itself until it is resumed in the foreground. Previously, the shell
+  would continue running in the background, interfering with the foreground
+  process group.
 - If job control is enabled and the shell does not have a controlling terminal,
   the shell now proceeds without managing foreground-ness of process groups.
   Jobs are still assigned to their own process groups. Previously, the shell